requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
black==23.10.1
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
fake-useragent==1.3.0
frozenlist==1.4.0
idna==3.4
lxml==4.9.3
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
packaging==23.2
pandas==2.1.4
pathspec==0.11.2
platformdirs==3.11.0
pycodestyle==2.11.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2023.4
urllib3==2.0.7
yarl==1.9.2
